@@ -8,43 +8,55 @@
     "type": "git",
     "url": "https://github.com/MatrixAI/js-logger.git"
   },
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "type": "module",
+  "exports": {
+    "./package.json": "./package.json",
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    },
+    "./*.js": {
+      "types": "./dist/*.d.ts",
+      "import": "./dist/*.js"
+    },
+    "./*": "./dist/*"
+  },
+  "imports": {
+    "#*": "./dist/*"
+  },
   "scripts": {
     "prepare": "tsc -p ./tsconfig.build.json",
     "build": "shx rm -rf ./dist && tsc -p ./tsconfig.build.json",
     "postversion": "npm install --package-lock-only --ignore-scripts --silent",
-    "ts-node": "ts-node",
-    "test": "jest",
-    "lint": "eslint '{src,tests,scripts,benches}/**/*.{js,ts}'",
-    "lintfix": "eslint '{src,tests,scripts,benches}/**/*.{js,ts}' --fix",
+    "tsx": "tsx",
+    "test": "node ./scripts/test.mjs",
+    "lint": "eslint '{src,tests,scripts,benches}/**/*.{js,mjs,ts,mts,jsx,tsx}'",
+    "lintfix": "eslint '{src,tests,scripts,benches}/**/*.{js,mjs,ts,mts,jsx,tsx}' --fix",
     "lint-shell": "find ./src ./tests ./scripts -type f -regextype posix-extended -regex '.*\\.(sh)' -exec shellcheck {} +",
     "docs": "shx rm -rf ./docs && typedoc --gitRevision master --tsconfig ./tsconfig.build.json --out ./docs src",
-    "bench": "shx rm -rf ./benches/results && ts-node ./benches"
+    "bench": "tsc -p ./tsconfig.build.json && shx rm -rf ./benches/results && tsx ./benches/index.ts"
   },
   "devDependencies": {
-    "@swc/core": "^1.3.62",
-    "@swc/jest": "^0.2.26",
-    "@types/jest": "^28.1.3",
+    "@swc/core": "^1.3.76",
+    "@swc/jest": "^0.2.29",
+    "@types/jest": "^29.5.2",
     "@types/node": "^18.15.0",
-    "@typescript-eslint/eslint-plugin": "^5.45.1",
-    "@typescript-eslint/parser": "^5.45.1",
+    "@typescript-eslint/eslint-plugin": "^5.61.0",
+    "@typescript-eslint/parser": "^5.61.0",
     "benny": "^3.7.1",
     "common-tags": "^1.8.2",
-    "eslint": "^8.15.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^28.1.1",
-    "jest-extended": "^3.0.1",
-    "jest-junit": "^14.0.0",
-    "prettier": "^2.6.2",
+    "eslint": "^8.44.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^5.0.0-alpha.2",
+    "jest": "^29.6.2",
+    "jest-extended": "^4.0.0",
+    "jest-junit": "^16.0.0",
+    "prettier": "^3.0.0",
     "shx": "^0.3.4",
     "systeminformation": "^5.18.5",
-    "ts-jest": "^28.0.5",
-    "ts-node": "^10.9.1",
-    "tsconfig-paths": "^3.9.0",
-    "typedoc": "^0.23.21",
-    "typescript": "^4.9.3"
+    "tsx": "^3.12.7",
+    "typedoc": "^0.24.8",
+    "typescript": "^5.1.6"
   }
 }
