@@ -1,19 +1,22 @@
-import type { LogRecord } from '@/types';
-import path from 'path';
+import type { LogRecord } from '#types.js';
+import path from 'node:path';
+import url from 'node:url';
 import b from 'benny';
+import { BenchHandler, suiteCommon } from './utils/index.js';
 import Logger, {
   LogLevel,
   formatting,
   levelToString,
   evalLogDataValue,
-} from '@';
-import { BenchHandler, suiteCommon } from './utils';
+} from '#index.js';
+
+const filePath = url.fileURLToPath(import.meta.url);
 
 async function main() {
   const msg = 'Hello World';
   const data = { foo: 'bar', bar: () => 'foo' };
   const summary = await b.suite(
-    path.basename(__filename, path.extname(__filename)),
+    path.basename(filePath, path.extname(filePath)),
     b.add('formatting default', () => {
       const logger = new Logger('root', LogLevel.NOTSET, [
         new BenchHandler(formatting.jsonFormatter),
@@ -82,8 +85,11 @@ async function main() {
   return summary;
 }
 
-if (require.main === module) {
-  void main();
+if (import.meta.url.startsWith('file:')) {
+  const modulePath = url.fileURLToPath(import.meta.url);
+  if (process.argv[1] === modulePath) {
+    void main();
+  }
 }
 
 export default main;
