@@ -1,1 +1,1 @@
-window.searchData = JSON.parse("{\"kinds\":{\"4\":\"Namespace\",\"8\":\"Enumeration\",\"16\":\"Enumeration Member\",\"32\":\"Variable\",\"64\":\"Function\",\"128\":\"Class\",\"256\":\"Interface\",\"512\":\"Constructor\",\"1024\":\"Property\",\"2048\":\"Method\",\"65536\":\"Type literal\",\"4194304\":\"Type alias\"},\"rows\":[{\"kind\":128,\"name\":\"Handler\",\"url\":\"classes/Handler.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Handler.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"Handler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/Handler.html#formatter\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/Handler.html#setFormatter\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/Handler.html#handle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/Handler.html#format\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/Handler.html#emit\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"Handler\"},{\"kind\":4,\"name\":\"formatting\",\"url\":\"modules/formatting.html\",\"classes\":\"tsd-kind-namespace\"},{\"kind\":32,\"name\":\"level\",\"url\":\"variables/formatting.level.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"key\",\"url\":\"variables/formatting.key.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"keys\",\"url\":\"variables/formatting.keys.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"date\",\"url\":\"variables/formatting.date.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"msg\",\"url\":\"variables/formatting.msg.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"stack\",\"url\":\"variables/formatting.stack.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"data\",\"url\":\"variables/formatting.data.html\",\"classes\":\"tsd-kind-variable tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":64,\"name\":\"format\",\"url\":\"functions/formatting.format.html\",\"classes\":\"tsd-kind-function tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":64,\"name\":\"formatter\",\"url\":\"functions/formatting.formatter.html\",\"classes\":\"tsd-kind-function tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":64,\"name\":\"jsonFormatter\",\"url\":\"functions/formatting.jsonFormatter.html\",\"classes\":\"tsd-kind-function tsd-parent-kind-namespace\",\"parent\":\"formatting\"},{\"kind\":128,\"name\":\"ConsoleErrHandler\",\"url\":\"classes/ConsoleErrHandler.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/ConsoleErrHandler.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/ConsoleErrHandler.html#emit\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/ConsoleErrHandler.html#formatter\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/ConsoleErrHandler.html#setFormatter\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/ConsoleErrHandler.html#handle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/ConsoleErrHandler.html#format\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":128,\"name\":\"ConsoleOutHandler\",\"url\":\"classes/ConsoleOutHandler.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/ConsoleOutHandler.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/ConsoleOutHandler.html#emit\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/ConsoleOutHandler.html#formatter\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/ConsoleOutHandler.html#setFormatter\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/ConsoleOutHandler.html#handle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/ConsoleOutHandler.html#format\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":128,\"name\":\"StreamHandler\",\"url\":\"classes/StreamHandler.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/StreamHandler.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/StreamHandler.html#emit\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"StreamHandler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/StreamHandler.html#formatter\",\"classes\":\"tsd-kind-property tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/StreamHandler.html#setFormatter\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/StreamHandler.html#handle\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/StreamHandler.html#format\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":64,\"name\":\"levelToString\",\"url\":\"functions/levelToString.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"evalLogDataValue\",\"url\":\"functions/evalLogDataValue.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":64,\"name\":\"evalLogData\",\"url\":\"functions/evalLogData.html\",\"classes\":\"tsd-kind-function\"},{\"kind\":32,\"name\":\"hasCaptureStackTrace\",\"url\":\"variables/hasCaptureStackTrace.html\",\"classes\":\"tsd-kind-variable\"},{\"kind\":32,\"name\":\"hasStackTraceLimit\",\"url\":\"variables/hasStackTraceLimit.html\",\"classes\":\"tsd-kind-variable\"},{\"kind\":8,\"name\":\"LogLevel\",\"url\":\"enums/LogLevel.html\",\"classes\":\"tsd-kind-enum\"},{\"kind\":16,\"name\":\"NOTSET\",\"url\":\"enums/LogLevel.html#NOTSET\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"DEBUG\",\"url\":\"enums/LogLevel.html#DEBUG\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"INFO\",\"url\":\"enums/LogLevel.html#INFO\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"WARN\",\"url\":\"enums/LogLevel.html#WARN\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"ERROR\",\"url\":\"enums/LogLevel.html#ERROR\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"SILENT\",\"url\":\"enums/LogLevel.html#SILENT\",\"classes\":\"tsd-kind-enum-member tsd-parent-kind-enum\",\"parent\":\"LogLevel\"},{\"kind\":256,\"name\":\"ToString\",\"url\":\"interfaces/ToString.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"toString\",\"url\":\"interfaces/ToString.html#toString\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ToString\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/ToString.html#toString.__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-property\",\"parent\":\"ToString.toString\"},{\"kind\":256,\"name\":\"ToJSON\",\"url\":\"interfaces/ToJSON.html\",\"classes\":\"tsd-kind-interface\"},{\"kind\":1024,\"name\":\"toJSON\",\"url\":\"interfaces/ToJSON.html#toJSON\",\"classes\":\"tsd-kind-property tsd-parent-kind-interface\",\"parent\":\"ToJSON\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/ToJSON.html#toJSON.__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-property\",\"parent\":\"ToJSON.toJSON\"},{\"kind\":4194304,\"name\":\"LogDataKey\",\"url\":\"types/LogDataKey.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"LogDataValue\",\"url\":\"types/LogDataValue.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"LogData\",\"url\":\"types/LogData.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":4194304,\"name\":\"LogRecord\",\"url\":\"types/LogRecord.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogRecord.html#__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-type-alias\",\"parent\":\"LogRecord\"},{\"kind\":1024,\"name\":\"logger\",\"url\":\"types/LogRecord.html#__type.logger\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"key\",\"url\":\"types/LogRecord.html#__type.key\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"keys\",\"url\":\"types/LogRecord.html#__type.keys\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"level\",\"url\":\"types/LogRecord.html#__type.level\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"msg\",\"url\":\"types/LogRecord.html#__type.msg\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"types/LogRecord.html#__type.data\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"date\",\"url\":\"types/LogRecord.html#__type.date\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogRecord.html#__type.date.__type-1\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-property\",\"parent\":\"LogRecord.__type.date\"},{\"kind\":1024,\"name\":\"stack\",\"url\":\"types/LogRecord.html#__type.stack\",\"classes\":\"tsd-kind-property tsd-parent-kind-type-literal\",\"parent\":\"LogRecord.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogRecord.html#__type.stack.__type-3\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-property\",\"parent\":\"LogRecord.__type.stack\"},{\"kind\":4194304,\"name\":\"LogFormatter\",\"url\":\"types/LogFormatter.html\",\"classes\":\"tsd-kind-type-alias\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogFormatter.html#__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-type-alias\",\"parent\":\"LogFormatter\"},{\"kind\":128,\"name\":\"default\",\"url\":\"classes/default.html\",\"classes\":\"tsd-kind-class\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/default.html#constructor\",\"classes\":\"tsd-kind-constructor tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"key\",\"url\":\"classes/default.html#key\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"level\",\"url\":\"classes/default.html#level\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"filter\",\"url\":\"classes/default.html#filter\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"keys\",\"url\":\"classes/default.html#keys\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"handlers\",\"url\":\"classes/default.html#handlers\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"parent\",\"url\":\"classes/default.html#parent\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"loggers\",\"url\":\"classes/default.html#loggers\",\"classes\":\"tsd-kind-property tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"classes/default.html#loggers.__type\",\"classes\":\"tsd-kind-type-literal tsd-parent-kind-property\",\"parent\":\"default.loggers\"},{\"kind\":2048,\"name\":\"getChild\",\"url\":\"classes/default.html#getChild\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"getParent\",\"url\":\"classes/default.html#getParent\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"setLevel\",\"url\":\"classes/default.html#setLevel\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"getEffectiveLevel\",\"url\":\"classes/default.html#getEffectiveLevel\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"isEnabledFor\",\"url\":\"classes/default.html#isEnabledFor\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"addHandler\",\"url\":\"classes/default.html#addHandler\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"removeHandler\",\"url\":\"classes/default.html#removeHandler\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"clearHandlers\",\"url\":\"classes/default.html#clearHandlers\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"hasHandlers\",\"url\":\"classes/default.html#hasHandlers\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"setFilter\",\"url\":\"classes/default.html#setFilter\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"unsetFilter\",\"url\":\"classes/default.html#unsetFilter\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"debug\",\"url\":\"classes/default.html#debug\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"info\",\"url\":\"classes/default.html#info\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"warn\",\"url\":\"classes/default.html#warn\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"error\",\"url\":\"classes/default.html#error\",\"classes\":\"tsd-kind-method tsd-parent-kind-class\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"log\",\"url\":\"classes/default.html#log\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"makeRecord\",\"url\":\"classes/default.html#makeRecord\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"callHandlers\",\"url\":\"classes/default.html#callHandlers\",\"classes\":\"tsd-kind-method tsd-parent-kind-class tsd-is-protected\",\"parent\":\"default\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,42.293]],[\"comment/0\",[]],[\"name/1\",[1,29.3]],[\"comment/1\",[]],[\"name/2\",[2,29.3]],[\"comment/2\",[]],[\"name/3\",[3,31.307]],[\"comment/3\",[]],[\"name/4\",[4,31.307]],[\"comment/4\",[]],[\"name/5\",[5,29.3]],[\"comment/5\",[]],[\"name/6\",[6,31.307]],[\"comment/6\",[]],[\"name/7\",[7,42.293]],[\"comment/7\",[]],[\"name/8\",[8,33.82]],[\"comment/8\",[]],[\"name/9\",[9,33.82]],[\"comment/9\",[]],[\"name/10\",[10,33.82]],[\"comment/10\",[]],[\"name/11\",[11,37.184]],[\"comment/11\",[]],[\"name/12\",[12,37.184]],[\"comment/12\",[]],[\"name/13\",[13,37.184]],[\"comment/13\",[]],[\"name/14\",[14,37.184]],[\"comment/14\",[]],[\"name/15\",[5,29.3]],[\"comment/15\",[]],[\"name/16\",[2,29.3]],[\"comment/16\",[]],[\"name/17\",[15,42.293]],[\"comment/17\",[]],[\"name/18\",[16,42.293]],[\"comment/18\",[]],[\"name/19\",[1,29.3]],[\"comment/19\",[]],[\"name/20\",[6,31.307]],[\"comment/20\",[]],[\"name/21\",[2,29.3]],[\"comment/21\",[]],[\"name/22\",[3,31.307]],[\"comment/22\",[]],[\"name/23\",[4,31.307]],[\"comment/23\",[]],[\"name/24\",[5,29.3]],[\"comment/24\",[]],[\"name/25\",[17,42.293]],[\"comment/25\",[]],[\"name/26\",[1,29.3]],[\"comment/26\",[]],[\"name/27\",[6,31.307]],[\"comment/27\",[]],[\"name/28\",[2,29.3]],[\"comment/28\",[]],[\"name/29\",[3,31.307]],[\"comment/29\",[]],[\"name/30\",[4,31.307]],[\"comment/30\",[]],[\"name/31\",[5,29.3]],[\"comment/31\",[]],[\"name/32\",[18,42.293]],[\"comment/32\",[]],[\"name/33\",[1,29.3]],[\"comment/33\",[]],[\"name/34\",[6,31.307]],[\"comment/34\",[]],[\"name/35\",[2,29.3]],[\"comment/35\",[]],[\"name/36\",[3,31.307]],[\"comment/36\",[]],[\"name/37\",[4,31.307]],[\"comment/37\",[]],[\"name/38\",[5,29.3]],[\"comment/38\",[]],[\"name/39\",[19,42.293]],[\"comment/39\",[]],[\"name/40\",[20,42.293]],[\"comment/40\",[]],[\"name/41\",[21,42.293]],[\"comment/41\",[]],[\"name/42\",[22,42.293]],[\"comment/42\",[]],[\"name/43\",[23,42.293]],[\"comment/43\",[]],[\"name/44\",[24,42.293]],[\"comment/44\",[]],[\"name/45\",[25,42.293]],[\"comment/45\",[]],[\"name/46\",[26,37.184]],[\"comment/46\",[]],[\"name/47\",[27,37.184]],[\"comment/47\",[]],[\"name/48\",[28,37.184]],[\"comment/48\",[]],[\"name/49\",[29,37.184]],[\"comment/49\",[]],[\"name/50\",[30,42.293]],[\"comment/50\",[]],[\"name/51\",[31,37.184]],[\"comment/51\",[]],[\"name/52\",[31,37.184]],[\"comment/52\",[]],[\"name/53\",[32,26.198]],[\"comment/53\",[]],[\"name/54\",[33,37.184]],[\"comment/54\",[]],[\"name/55\",[33,37.184]],[\"comment/55\",[]],[\"name/56\",[32,26.198]],[\"comment/56\",[]],[\"name/57\",[34,42.293]],[\"comment/57\",[]],[\"name/58\",[35,42.293]],[\"comment/58\",[]],[\"name/59\",[36,42.293]],[\"comment/59\",[]],[\"name/60\",[37,42.293]],[\"comment/60\",[]],[\"name/61\",[32,26.198]],[\"comment/61\",[]],[\"name/62\",[38,42.293]],[\"comment/62\",[]],[\"name/63\",[9,33.82]],[\"comment/63\",[]],[\"name/64\",[10,33.82]],[\"comment/64\",[]],[\"name/65\",[8,33.82]],[\"comment/65\",[]],[\"name/66\",[12,37.184]],[\"comment/66\",[]],[\"name/67\",[14,37.184]],[\"comment/67\",[]],[\"name/68\",[11,37.184]],[\"comment/68\",[]],[\"name/69\",[32,26.198]],[\"comment/69\",[]],[\"name/70\",[13,37.184]],[\"comment/70\",[]],[\"name/71\",[32,26.198]],[\"comment/71\",[]],[\"name/72\",[39,42.293]],[\"comment/72\",[]],[\"name/73\",[32,26.198]],[\"comment/73\",[]],[\"name/74\",[40,42.293]],[\"comment/74\",[]],[\"name/75\",[1,29.3]],[\"comment/75\",[]],[\"name/76\",[9,33.82]],[\"comment/76\",[]],[\"name/77\",[8,33.82]],[\"comment/77\",[]],[\"name/78\",[41,42.293]],[\"comment/78\",[]],[\"name/79\",[10,33.82]],[\"comment/79\",[]],[\"name/80\",[42,42.293]],[\"comment/80\",[]],[\"name/81\",[43,42.293]],[\"comment/81\",[]],[\"name/82\",[44,42.293]],[\"comment/82\",[]],[\"name/83\",[32,26.198]],[\"comment/83\",[]],[\"name/84\",[45,42.293]],[\"comment/84\",[]],[\"name/85\",[46,42.293]],[\"comment/85\",[]],[\"name/86\",[47,42.293]],[\"comment/86\",[]],[\"name/87\",[48,42.293]],[\"comment/87\",[]],[\"name/88\",[49,42.293]],[\"comment/88\",[]],[\"name/89\",[50,42.293]],[\"comment/89\",[]],[\"name/90\",[51,42.293]],[\"comment/90\",[]],[\"name/91\",[52,42.293]],[\"comment/91\",[]],[\"name/92\",[53,42.293]],[\"comment/92\",[]],[\"name/93\",[54,42.293]],[\"comment/93\",[]],[\"name/94\",[55,42.293]],[\"comment/94\",[]],[\"name/95\",[26,37.184]],[\"comment/95\",[]],[\"name/96\",[27,37.184]],[\"comment/96\",[]],[\"name/97\",[28,37.184]],[\"comment/97\",[]],[\"name/98\",[29,37.184]],[\"comment/98\",[]],[\"name/99\",[56,42.293]],[\"comment/99\",[]],[\"name/100\",[57,42.293]],[\"comment/100\",[]],[\"name/101\",[58,42.293]],[\"comment/101\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":32,\"name\":{\"53\":{},\"56\":{},\"61\":{},\"69\":{},\"71\":{},\"73\":{},\"83\":{}},\"comment\":{}}],[\"addhandler\",{\"_index\":50,\"name\":{\"89\":{}},\"comment\":{}}],[\"callhandlers\",{\"_index\":58,\"name\":{\"101\":{}},\"comment\":{}}],[\"clearhandlers\",{\"_index\":52,\"name\":{\"91\":{}},\"comment\":{}}],[\"consoleerrhandler\",{\"_index\":16,\"name\":{\"18\":{}},\"comment\":{}}],[\"consoleouthandler\",{\"_index\":17,\"name\":{\"25\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":1,\"name\":{\"1\":{},\"19\":{},\"26\":{},\"33\":{},\"75\":{}},\"comment\":{}}],[\"data\",{\"_index\":14,\"name\":{\"14\":{},\"67\":{}},\"comment\":{}}],[\"date\",{\"_index\":11,\"name\":{\"11\":{},\"68\":{}},\"comment\":{}}],[\"debug\",{\"_index\":26,\"name\":{\"46\":{},\"95\":{}},\"comment\":{}}],[\"default\",{\"_index\":40,\"name\":{\"74\":{}},\"comment\":{}}],[\"emit\",{\"_index\":6,\"name\":{\"6\":{},\"20\":{},\"27\":{},\"34\":{}},\"comment\":{}}],[\"error\",{\"_index\":29,\"name\":{\"49\":{},\"98\":{}},\"comment\":{}}],[\"evallogdata\",{\"_index\":21,\"name\":{\"41\":{}},\"comment\":{}}],[\"evallogdatavalue\",{\"_index\":20,\"name\":{\"40\":{}},\"comment\":{}}],[\"filter\",{\"_index\":41,\"name\":{\"78\":{}},\"comment\":{}}],[\"format\",{\"_index\":5,\"name\":{\"5\":{},\"15\":{},\"24\":{},\"31\":{},\"38\":{}},\"comment\":{}}],[\"formatter\",{\"_index\":2,\"name\":{\"2\":{},\"16\":{},\"21\":{},\"28\":{},\"35\":{}},\"comment\":{}}],[\"formatting\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"getchild\",{\"_index\":45,\"name\":{\"84\":{}},\"comment\":{}}],[\"geteffectivelevel\",{\"_index\":48,\"name\":{\"87\":{}},\"comment\":{}}],[\"getparent\",{\"_index\":46,\"name\":{\"85\":{}},\"comment\":{}}],[\"handle\",{\"_index\":4,\"name\":{\"4\":{},\"23\":{},\"30\":{},\"37\":{}},\"comment\":{}}],[\"handler\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"handlers\",{\"_index\":42,\"name\":{\"80\":{}},\"comment\":{}}],[\"hascapturestacktrace\",{\"_index\":22,\"name\":{\"42\":{}},\"comment\":{}}],[\"hashandlers\",{\"_index\":53,\"name\":{\"92\":{}},\"comment\":{}}],[\"hasstacktracelimit\",{\"_index\":23,\"name\":{\"43\":{}},\"comment\":{}}],[\"info\",{\"_index\":27,\"name\":{\"47\":{},\"96\":{}},\"comment\":{}}],[\"isenabledfor\",{\"_index\":49,\"name\":{\"88\":{}},\"comment\":{}}],[\"jsonformatter\",{\"_index\":15,\"name\":{\"17\":{}},\"comment\":{}}],[\"key\",{\"_index\":9,\"name\":{\"9\":{},\"63\":{},\"76\":{}},\"comment\":{}}],[\"keys\",{\"_index\":10,\"name\":{\"10\":{},\"64\":{},\"79\":{}},\"comment\":{}}],[\"level\",{\"_index\":8,\"name\":{\"8\":{},\"65\":{},\"77\":{}},\"comment\":{}}],[\"leveltostring\",{\"_index\":19,\"name\":{\"39\":{}},\"comment\":{}}],[\"log\",{\"_index\":56,\"name\":{\"99\":{}},\"comment\":{}}],[\"logdata\",{\"_index\":36,\"name\":{\"59\":{}},\"comment\":{}}],[\"logdatakey\",{\"_index\":34,\"name\":{\"57\":{}},\"comment\":{}}],[\"logdatavalue\",{\"_index\":35,\"name\":{\"58\":{}},\"comment\":{}}],[\"logformatter\",{\"_index\":39,\"name\":{\"72\":{}},\"comment\":{}}],[\"logger\",{\"_index\":38,\"name\":{\"62\":{}},\"comment\":{}}],[\"loggers\",{\"_index\":44,\"name\":{\"82\":{}},\"comment\":{}}],[\"loglevel\",{\"_index\":24,\"name\":{\"44\":{}},\"comment\":{}}],[\"logrecord\",{\"_index\":37,\"name\":{\"60\":{}},\"comment\":{}}],[\"makerecord\",{\"_index\":57,\"name\":{\"100\":{}},\"comment\":{}}],[\"msg\",{\"_index\":12,\"name\":{\"12\":{},\"66\":{}},\"comment\":{}}],[\"notset\",{\"_index\":25,\"name\":{\"45\":{}},\"comment\":{}}],[\"parent\",{\"_index\":43,\"name\":{\"81\":{}},\"comment\":{}}],[\"removehandler\",{\"_index\":51,\"name\":{\"90\":{}},\"comment\":{}}],[\"setfilter\",{\"_index\":54,\"name\":{\"93\":{}},\"comment\":{}}],[\"setformatter\",{\"_index\":3,\"name\":{\"3\":{},\"22\":{},\"29\":{},\"36\":{}},\"comment\":{}}],[\"setlevel\",{\"_index\":47,\"name\":{\"86\":{}},\"comment\":{}}],[\"silent\",{\"_index\":30,\"name\":{\"50\":{}},\"comment\":{}}],[\"stack\",{\"_index\":13,\"name\":{\"13\":{},\"70\":{}},\"comment\":{}}],[\"streamhandler\",{\"_index\":18,\"name\":{\"32\":{}},\"comment\":{}}],[\"tojson\",{\"_index\":33,\"name\":{\"54\":{},\"55\":{}},\"comment\":{}}],[\"tostring\",{\"_index\":31,\"name\":{\"51\":{},\"52\":{}},\"comment\":{}}],[\"unsetfilter\",{\"_index\":55,\"name\":{\"94\":{}},\"comment\":{}}],[\"warn\",{\"_index\":28,\"name\":{\"48\":{},\"97\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
+window.searchData = JSON.parse("{\"rows\":[{\"kind\":128,\"name\":\"Handler\",\"url\":\"classes/Handler.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/Handler.html#constructor\",\"classes\":\"\",\"parent\":\"Handler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/Handler.html#formatter\",\"classes\":\"\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/Handler.html#setFormatter\",\"classes\":\"\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/Handler.html#handle\",\"classes\":\"\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/Handler.html#format\",\"classes\":\"\",\"parent\":\"Handler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/Handler.html#emit\",\"classes\":\"\",\"parent\":\"Handler\"},{\"kind\":4,\"name\":\"formatting\",\"url\":\"modules/formatting.html\",\"classes\":\"\"},{\"kind\":32,\"name\":\"level\",\"url\":\"variables/formatting.level.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"key\",\"url\":\"variables/formatting.key.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"keys\",\"url\":\"variables/formatting.keys.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"date\",\"url\":\"variables/formatting.date.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"msg\",\"url\":\"variables/formatting.msg.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"stack\",\"url\":\"variables/formatting.stack.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":32,\"name\":\"data\",\"url\":\"variables/formatting.data.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":64,\"name\":\"format\",\"url\":\"functions/formatting.format.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":64,\"name\":\"formatter\",\"url\":\"functions/formatting.formatter.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":64,\"name\":\"jsonFormatter\",\"url\":\"functions/formatting.jsonFormatter.html\",\"classes\":\"\",\"parent\":\"formatting\"},{\"kind\":128,\"name\":\"ConsoleErrHandler\",\"url\":\"classes/ConsoleErrHandler.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/ConsoleErrHandler.html#constructor\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/ConsoleErrHandler.html#emit\",\"classes\":\"\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/ConsoleErrHandler.html#formatter\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/ConsoleErrHandler.html#setFormatter\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/ConsoleErrHandler.html#handle\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/ConsoleErrHandler.html#format\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleErrHandler\"},{\"kind\":128,\"name\":\"ConsoleOutHandler\",\"url\":\"classes/ConsoleOutHandler.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/ConsoleOutHandler.html#constructor\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/ConsoleOutHandler.html#emit\",\"classes\":\"\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/ConsoleOutHandler.html#formatter\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/ConsoleOutHandler.html#setFormatter\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/ConsoleOutHandler.html#handle\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/ConsoleOutHandler.html#format\",\"classes\":\"tsd-is-inherited\",\"parent\":\"ConsoleOutHandler\"},{\"kind\":128,\"name\":\"StreamHandler\",\"url\":\"classes/StreamHandler.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/StreamHandler.html#constructor\",\"classes\":\"tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"emit\",\"url\":\"classes/StreamHandler.html#emit\",\"classes\":\"\",\"parent\":\"StreamHandler\"},{\"kind\":1024,\"name\":\"formatter\",\"url\":\"classes/StreamHandler.html#formatter\",\"classes\":\"tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"setFormatter\",\"url\":\"classes/StreamHandler.html#setFormatter\",\"classes\":\"tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"handle\",\"url\":\"classes/StreamHandler.html#handle\",\"classes\":\"tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":2048,\"name\":\"format\",\"url\":\"classes/StreamHandler.html#format\",\"classes\":\"tsd-is-inherited\",\"parent\":\"StreamHandler\"},{\"kind\":64,\"name\":\"levelToString\",\"url\":\"functions/levelToString.html\",\"classes\":\"\"},{\"kind\":64,\"name\":\"evalLogDataValue\",\"url\":\"functions/evalLogDataValue.html\",\"classes\":\"\"},{\"kind\":64,\"name\":\"evalLogData\",\"url\":\"functions/evalLogData.html\",\"classes\":\"\"},{\"kind\":32,\"name\":\"hasCaptureStackTrace\",\"url\":\"variables/hasCaptureStackTrace.html\",\"classes\":\"\"},{\"kind\":32,\"name\":\"hasStackTraceLimit\",\"url\":\"variables/hasStackTraceLimit.html\",\"classes\":\"\"},{\"kind\":8,\"name\":\"LogLevel\",\"url\":\"enums/LogLevel.html\",\"classes\":\"\"},{\"kind\":16,\"name\":\"NOTSET\",\"url\":\"enums/LogLevel.html#NOTSET\",\"classes\":\"\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"DEBUG\",\"url\":\"enums/LogLevel.html#DEBUG\",\"classes\":\"\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"INFO\",\"url\":\"enums/LogLevel.html#INFO\",\"classes\":\"\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"WARN\",\"url\":\"enums/LogLevel.html#WARN\",\"classes\":\"\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"ERROR\",\"url\":\"enums/LogLevel.html#ERROR\",\"classes\":\"\",\"parent\":\"LogLevel\"},{\"kind\":16,\"name\":\"SILENT\",\"url\":\"enums/LogLevel.html#SILENT\",\"classes\":\"\",\"parent\":\"LogLevel\"},{\"kind\":256,\"name\":\"ToString\",\"url\":\"interfaces/ToString.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"toString\",\"url\":\"interfaces/ToString.html#toString\",\"classes\":\"\",\"parent\":\"ToString\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/ToString.html#toString.__type\",\"classes\":\"\",\"parent\":\"ToString.toString\"},{\"kind\":256,\"name\":\"ToJSON\",\"url\":\"interfaces/ToJSON.html\",\"classes\":\"\"},{\"kind\":1024,\"name\":\"toJSON\",\"url\":\"interfaces/ToJSON.html#toJSON\",\"classes\":\"\",\"parent\":\"ToJSON\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"interfaces/ToJSON.html#toJSON.__type\",\"classes\":\"\",\"parent\":\"ToJSON.toJSON\"},{\"kind\":4194304,\"name\":\"LogDataKey\",\"url\":\"types/LogDataKey.html\",\"classes\":\"\"},{\"kind\":4194304,\"name\":\"LogDataValue\",\"url\":\"types/LogDataValue.html\",\"classes\":\"\"},{\"kind\":4194304,\"name\":\"LogData\",\"url\":\"types/LogData.html\",\"classes\":\"\"},{\"kind\":4194304,\"name\":\"LogRecord\",\"url\":\"types/LogRecord.html\",\"classes\":\"\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogRecord.html#__type\",\"classes\":\"\",\"parent\":\"LogRecord\"},{\"kind\":1024,\"name\":\"logger\",\"url\":\"types/LogRecord.html#__type.logger\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"key\",\"url\":\"types/LogRecord.html#__type.key\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"keys\",\"url\":\"types/LogRecord.html#__type.keys\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"level\",\"url\":\"types/LogRecord.html#__type.level\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"msg\",\"url\":\"types/LogRecord.html#__type.msg\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"data\",\"url\":\"types/LogRecord.html#__type.data\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":1024,\"name\":\"date\",\"url\":\"types/LogRecord.html#__type.date\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogRecord.html#__type.date.__type-1\",\"classes\":\"\",\"parent\":\"LogRecord.__type.date\"},{\"kind\":1024,\"name\":\"stack\",\"url\":\"types/LogRecord.html#__type.stack\",\"classes\":\"\",\"parent\":\"LogRecord.__type\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogRecord.html#__type.stack.__type-3\",\"classes\":\"\",\"parent\":\"LogRecord.__type.stack\"},{\"kind\":4194304,\"name\":\"LogFormatter\",\"url\":\"types/LogFormatter.html\",\"classes\":\"\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"types/LogFormatter.html#__type\",\"classes\":\"\",\"parent\":\"LogFormatter\"},{\"kind\":128,\"name\":\"default\",\"url\":\"classes/default.html\",\"classes\":\"\"},{\"kind\":512,\"name\":\"constructor\",\"url\":\"classes/default.html#constructor\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"key\",\"url\":\"classes/default.html#key\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"level\",\"url\":\"classes/default.html#level\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"filter\",\"url\":\"classes/default.html#filter\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"keys\",\"url\":\"classes/default.html#keys\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"handlers\",\"url\":\"classes/default.html#handlers\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"parent\",\"url\":\"classes/default.html#parent\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":1024,\"name\":\"loggers\",\"url\":\"classes/default.html#loggers\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":65536,\"name\":\"__type\",\"url\":\"classes/default.html#loggers.__type\",\"classes\":\"\",\"parent\":\"default.loggers\"},{\"kind\":2048,\"name\":\"getChild\",\"url\":\"classes/default.html#getChild\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"getParent\",\"url\":\"classes/default.html#getParent\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"setLevel\",\"url\":\"classes/default.html#setLevel\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"getEffectiveLevel\",\"url\":\"classes/default.html#getEffectiveLevel\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"isEnabledFor\",\"url\":\"classes/default.html#isEnabledFor\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"addHandler\",\"url\":\"classes/default.html#addHandler\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"removeHandler\",\"url\":\"classes/default.html#removeHandler\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"clearHandlers\",\"url\":\"classes/default.html#clearHandlers\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"hasHandlers\",\"url\":\"classes/default.html#hasHandlers\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"setFilter\",\"url\":\"classes/default.html#setFilter\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"unsetFilter\",\"url\":\"classes/default.html#unsetFilter\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"debug\",\"url\":\"classes/default.html#debug\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"info\",\"url\":\"classes/default.html#info\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"warn\",\"url\":\"classes/default.html#warn\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"error\",\"url\":\"classes/default.html#error\",\"classes\":\"\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"log\",\"url\":\"classes/default.html#log\",\"classes\":\"tsd-is-protected\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"makeRecord\",\"url\":\"classes/default.html#makeRecord\",\"classes\":\"tsd-is-protected\",\"parent\":\"default\"},{\"kind\":2048,\"name\":\"callHandlers\",\"url\":\"classes/default.html#callHandlers\",\"classes\":\"tsd-is-protected\",\"parent\":\"default\"}],\"index\":{\"version\":\"2.3.9\",\"fields\":[\"name\",\"comment\"],\"fieldVectors\":[[\"name/0\",[0,42.293]],[\"comment/0\",[]],[\"name/1\",[1,29.3]],[\"comment/1\",[]],[\"name/2\",[2,29.3]],[\"comment/2\",[]],[\"name/3\",[3,31.307]],[\"comment/3\",[]],[\"name/4\",[4,31.307]],[\"comment/4\",[]],[\"name/5\",[5,29.3]],[\"comment/5\",[]],[\"name/6\",[6,31.307]],[\"comment/6\",[]],[\"name/7\",[7,42.293]],[\"comment/7\",[]],[\"name/8\",[8,33.82]],[\"comment/8\",[]],[\"name/9\",[9,33.82]],[\"comment/9\",[]],[\"name/10\",[10,33.82]],[\"comment/10\",[]],[\"name/11\",[11,37.184]],[\"comment/11\",[]],[\"name/12\",[12,37.184]],[\"comment/12\",[]],[\"name/13\",[13,37.184]],[\"comment/13\",[]],[\"name/14\",[14,37.184]],[\"comment/14\",[]],[\"name/15\",[5,29.3]],[\"comment/15\",[]],[\"name/16\",[2,29.3]],[\"comment/16\",[]],[\"name/17\",[15,42.293]],[\"comment/17\",[]],[\"name/18\",[16,42.293]],[\"comment/18\",[]],[\"name/19\",[1,29.3]],[\"comment/19\",[]],[\"name/20\",[6,31.307]],[\"comment/20\",[]],[\"name/21\",[2,29.3]],[\"comment/21\",[]],[\"name/22\",[3,31.307]],[\"comment/22\",[]],[\"name/23\",[4,31.307]],[\"comment/23\",[]],[\"name/24\",[5,29.3]],[\"comment/24\",[]],[\"name/25\",[17,42.293]],[\"comment/25\",[]],[\"name/26\",[1,29.3]],[\"comment/26\",[]],[\"name/27\",[6,31.307]],[\"comment/27\",[]],[\"name/28\",[2,29.3]],[\"comment/28\",[]],[\"name/29\",[3,31.307]],[\"comment/29\",[]],[\"name/30\",[4,31.307]],[\"comment/30\",[]],[\"name/31\",[5,29.3]],[\"comment/31\",[]],[\"name/32\",[18,42.293]],[\"comment/32\",[]],[\"name/33\",[1,29.3]],[\"comment/33\",[]],[\"name/34\",[6,31.307]],[\"comment/34\",[]],[\"name/35\",[2,29.3]],[\"comment/35\",[]],[\"name/36\",[3,31.307]],[\"comment/36\",[]],[\"name/37\",[4,31.307]],[\"comment/37\",[]],[\"name/38\",[5,29.3]],[\"comment/38\",[]],[\"name/39\",[19,42.293]],[\"comment/39\",[]],[\"name/40\",[20,42.293]],[\"comment/40\",[]],[\"name/41\",[21,42.293]],[\"comment/41\",[]],[\"name/42\",[22,42.293]],[\"comment/42\",[]],[\"name/43\",[23,42.293]],[\"comment/43\",[]],[\"name/44\",[24,42.293]],[\"comment/44\",[]],[\"name/45\",[25,42.293]],[\"comment/45\",[]],[\"name/46\",[26,37.184]],[\"comment/46\",[]],[\"name/47\",[27,37.184]],[\"comment/47\",[]],[\"name/48\",[28,37.184]],[\"comment/48\",[]],[\"name/49\",[29,37.184]],[\"comment/49\",[]],[\"name/50\",[30,42.293]],[\"comment/50\",[]],[\"name/51\",[31,37.184]],[\"comment/51\",[]],[\"name/52\",[31,37.184]],[\"comment/52\",[]],[\"name/53\",[32,26.198]],[\"comment/53\",[]],[\"name/54\",[33,37.184]],[\"comment/54\",[]],[\"name/55\",[33,37.184]],[\"comment/55\",[]],[\"name/56\",[32,26.198]],[\"comment/56\",[]],[\"name/57\",[34,42.293]],[\"comment/57\",[]],[\"name/58\",[35,42.293]],[\"comment/58\",[]],[\"name/59\",[36,42.293]],[\"comment/59\",[]],[\"name/60\",[37,42.293]],[\"comment/60\",[]],[\"name/61\",[32,26.198]],[\"comment/61\",[]],[\"name/62\",[38,42.293]],[\"comment/62\",[]],[\"name/63\",[9,33.82]],[\"comment/63\",[]],[\"name/64\",[10,33.82]],[\"comment/64\",[]],[\"name/65\",[8,33.82]],[\"comment/65\",[]],[\"name/66\",[12,37.184]],[\"comment/66\",[]],[\"name/67\",[14,37.184]],[\"comment/67\",[]],[\"name/68\",[11,37.184]],[\"comment/68\",[]],[\"name/69\",[32,26.198]],[\"comment/69\",[]],[\"name/70\",[13,37.184]],[\"comment/70\",[]],[\"name/71\",[32,26.198]],[\"comment/71\",[]],[\"name/72\",[39,42.293]],[\"comment/72\",[]],[\"name/73\",[32,26.198]],[\"comment/73\",[]],[\"name/74\",[40,42.293]],[\"comment/74\",[]],[\"name/75\",[1,29.3]],[\"comment/75\",[]],[\"name/76\",[9,33.82]],[\"comment/76\",[]],[\"name/77\",[8,33.82]],[\"comment/77\",[]],[\"name/78\",[41,42.293]],[\"comment/78\",[]],[\"name/79\",[10,33.82]],[\"comment/79\",[]],[\"name/80\",[42,42.293]],[\"comment/80\",[]],[\"name/81\",[43,42.293]],[\"comment/81\",[]],[\"name/82\",[44,42.293]],[\"comment/82\",[]],[\"name/83\",[32,26.198]],[\"comment/83\",[]],[\"name/84\",[45,42.293]],[\"comment/84\",[]],[\"name/85\",[46,42.293]],[\"comment/85\",[]],[\"name/86\",[47,42.293]],[\"comment/86\",[]],[\"name/87\",[48,42.293]],[\"comment/87\",[]],[\"name/88\",[49,42.293]],[\"comment/88\",[]],[\"name/89\",[50,42.293]],[\"comment/89\",[]],[\"name/90\",[51,42.293]],[\"comment/90\",[]],[\"name/91\",[52,42.293]],[\"comment/91\",[]],[\"name/92\",[53,42.293]],[\"comment/92\",[]],[\"name/93\",[54,42.293]],[\"comment/93\",[]],[\"name/94\",[55,42.293]],[\"comment/94\",[]],[\"name/95\",[26,37.184]],[\"comment/95\",[]],[\"name/96\",[27,37.184]],[\"comment/96\",[]],[\"name/97\",[28,37.184]],[\"comment/97\",[]],[\"name/98\",[29,37.184]],[\"comment/98\",[]],[\"name/99\",[56,42.293]],[\"comment/99\",[]],[\"name/100\",[57,42.293]],[\"comment/100\",[]],[\"name/101\",[58,42.293]],[\"comment/101\",[]]],\"invertedIndex\":[[\"__type\",{\"_index\":32,\"name\":{\"53\":{},\"56\":{},\"61\":{},\"69\":{},\"71\":{},\"73\":{},\"83\":{}},\"comment\":{}}],[\"addhandler\",{\"_index\":50,\"name\":{\"89\":{}},\"comment\":{}}],[\"callhandlers\",{\"_index\":58,\"name\":{\"101\":{}},\"comment\":{}}],[\"clearhandlers\",{\"_index\":52,\"name\":{\"91\":{}},\"comment\":{}}],[\"consoleerrhandler\",{\"_index\":16,\"name\":{\"18\":{}},\"comment\":{}}],[\"consoleouthandler\",{\"_index\":17,\"name\":{\"25\":{}},\"comment\":{}}],[\"constructor\",{\"_index\":1,\"name\":{\"1\":{},\"19\":{},\"26\":{},\"33\":{},\"75\":{}},\"comment\":{}}],[\"data\",{\"_index\":14,\"name\":{\"14\":{},\"67\":{}},\"comment\":{}}],[\"date\",{\"_index\":11,\"name\":{\"11\":{},\"68\":{}},\"comment\":{}}],[\"debug\",{\"_index\":26,\"name\":{\"46\":{},\"95\":{}},\"comment\":{}}],[\"default\",{\"_index\":40,\"name\":{\"74\":{}},\"comment\":{}}],[\"emit\",{\"_index\":6,\"name\":{\"6\":{},\"20\":{},\"27\":{},\"34\":{}},\"comment\":{}}],[\"error\",{\"_index\":29,\"name\":{\"49\":{},\"98\":{}},\"comment\":{}}],[\"evallogdata\",{\"_index\":21,\"name\":{\"41\":{}},\"comment\":{}}],[\"evallogdatavalue\",{\"_index\":20,\"name\":{\"40\":{}},\"comment\":{}}],[\"filter\",{\"_index\":41,\"name\":{\"78\":{}},\"comment\":{}}],[\"format\",{\"_index\":5,\"name\":{\"5\":{},\"15\":{},\"24\":{},\"31\":{},\"38\":{}},\"comment\":{}}],[\"formatter\",{\"_index\":2,\"name\":{\"2\":{},\"16\":{},\"21\":{},\"28\":{},\"35\":{}},\"comment\":{}}],[\"formatting\",{\"_index\":7,\"name\":{\"7\":{}},\"comment\":{}}],[\"getchild\",{\"_index\":45,\"name\":{\"84\":{}},\"comment\":{}}],[\"geteffectivelevel\",{\"_index\":48,\"name\":{\"87\":{}},\"comment\":{}}],[\"getparent\",{\"_index\":46,\"name\":{\"85\":{}},\"comment\":{}}],[\"handle\",{\"_index\":4,\"name\":{\"4\":{},\"23\":{},\"30\":{},\"37\":{}},\"comment\":{}}],[\"handler\",{\"_index\":0,\"name\":{\"0\":{}},\"comment\":{}}],[\"handlers\",{\"_index\":42,\"name\":{\"80\":{}},\"comment\":{}}],[\"hascapturestacktrace\",{\"_index\":22,\"name\":{\"42\":{}},\"comment\":{}}],[\"hashandlers\",{\"_index\":53,\"name\":{\"92\":{}},\"comment\":{}}],[\"hasstacktracelimit\",{\"_index\":23,\"name\":{\"43\":{}},\"comment\":{}}],[\"info\",{\"_index\":27,\"name\":{\"47\":{},\"96\":{}},\"comment\":{}}],[\"isenabledfor\",{\"_index\":49,\"name\":{\"88\":{}},\"comment\":{}}],[\"jsonformatter\",{\"_index\":15,\"name\":{\"17\":{}},\"comment\":{}}],[\"key\",{\"_index\":9,\"name\":{\"9\":{},\"63\":{},\"76\":{}},\"comment\":{}}],[\"keys\",{\"_index\":10,\"name\":{\"10\":{},\"64\":{},\"79\":{}},\"comment\":{}}],[\"level\",{\"_index\":8,\"name\":{\"8\":{},\"65\":{},\"77\":{}},\"comment\":{}}],[\"leveltostring\",{\"_index\":19,\"name\":{\"39\":{}},\"comment\":{}}],[\"log\",{\"_index\":56,\"name\":{\"99\":{}},\"comment\":{}}],[\"logdata\",{\"_index\":36,\"name\":{\"59\":{}},\"comment\":{}}],[\"logdatakey\",{\"_index\":34,\"name\":{\"57\":{}},\"comment\":{}}],[\"logdatavalue\",{\"_index\":35,\"name\":{\"58\":{}},\"comment\":{}}],[\"logformatter\",{\"_index\":39,\"name\":{\"72\":{}},\"comment\":{}}],[\"logger\",{\"_index\":38,\"name\":{\"62\":{}},\"comment\":{}}],[\"loggers\",{\"_index\":44,\"name\":{\"82\":{}},\"comment\":{}}],[\"loglevel\",{\"_index\":24,\"name\":{\"44\":{}},\"comment\":{}}],[\"logrecord\",{\"_index\":37,\"name\":{\"60\":{}},\"comment\":{}}],[\"makerecord\",{\"_index\":57,\"name\":{\"100\":{}},\"comment\":{}}],[\"msg\",{\"_index\":12,\"name\":{\"12\":{},\"66\":{}},\"comment\":{}}],[\"notset\",{\"_index\":25,\"name\":{\"45\":{}},\"comment\":{}}],[\"parent\",{\"_index\":43,\"name\":{\"81\":{}},\"comment\":{}}],[\"removehandler\",{\"_index\":51,\"name\":{\"90\":{}},\"comment\":{}}],[\"setfilter\",{\"_index\":54,\"name\":{\"93\":{}},\"comment\":{}}],[\"setformatter\",{\"_index\":3,\"name\":{\"3\":{},\"22\":{},\"29\":{},\"36\":{}},\"comment\":{}}],[\"setlevel\",{\"_index\":47,\"name\":{\"86\":{}},\"comment\":{}}],[\"silent\",{\"_index\":30,\"name\":{\"50\":{}},\"comment\":{}}],[\"stack\",{\"_index\":13,\"name\":{\"13\":{},\"70\":{}},\"comment\":{}}],[\"streamhandler\",{\"_index\":18,\"name\":{\"32\":{}},\"comment\":{}}],[\"tojson\",{\"_index\":33,\"name\":{\"54\":{},\"55\":{}},\"comment\":{}}],[\"tostring\",{\"_index\":31,\"name\":{\"51\":{},\"52\":{}},\"comment\":{}}],[\"unsetfilter\",{\"_index\":55,\"name\":{\"94\":{}},\"comment\":{}}],[\"warn\",{\"_index\":28,\"name\":{\"48\":{},\"97\":{}},\"comment\":{}}]],\"pipeline\":[]}}");
